// app/interfaces/interfaces.d.ts
export interface ITimeFilterConfigItem {
  layerView: __esri.LayerView;
  increments:
    | "milliseconds"
    | "seconds"
    | "minutes"
    | "hours"
    | "days"
    | "weeks"
    | "months"
    | "years"
    | "decades"
    | "centuries";
  rangeEnd: Date;
  rangeStart: Date;
  currentTimeExtent: __esri.TimeExtent;
}